+++ public/userscripts/manga-tracker.user.js
@@ -7,7 +7,7 @@
 // @supportURL   https://github.com/DakuTree/manga-tracker/issues
 // @icon         https://trackr.moe/favicon.production.png
 // @include      /^https:\/\/(?:(?:dev|test)\.)?trackr\.moe(\/.*$|$)/
-// @include      /^https?:\/\/mangafox\.me\/manga\/.+\/(?:.*\/)?.*\/.*$/
+// @include      /^https?:\/\/mangafox\.(?:me|la)\/manga\/.+\/(?:.*\/)?.*\/.*$/
 // @include      /^https?:\/\/(?:www\.)?mangahere\.co\/manga\/.+\/.*\/?.*\/.*$/
 // @include      /^https?:\/\/bato\.to\/reader.*$/
 // @include      /^https:/\/dynasty-scans\.com\/chapters\/.+$/
@@ -45,8 +45,8 @@
 // @include      /^http:\/\/ravens-scans\.com\/(?:multi|lector)\/read\/.*?\/[a-z]+\/[0-9]+\/[0-9\.]+(\/.*)?$/
 // @include      /^https?:\/\/reader\.thecatscans\.com\/read\/.*?\/[a-z]+\/[0-9]+\/[0-9]+(\/.*)?$/
 // @include      /^http:\/\/hatigarmscans\.eu\/hs\/read\/.*?\/[a-z]+\/[0-9]+\/[0-9]+(\/.*)?$/
-// @updated      2017-12-03
-// @version      1.8.5
+// @updated      2017-12-06
+// @version      1.8.6
 // @downloadURL  https://trackr.moe/userscripts/manga-tracker.user.js
 // @updateURL    https://trackr.moe/userscripts/manga-tracker.meta.js
 // @require      https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js
@@ -64,6 +64,7 @@
 // @noframes
 // @connect      myanimelist.net
 // @connect      m.mangafox.me
+// @connect      m.mangafox.la
 // @connect      m.mangahere.co
 // @run-at       document-start
 // ==/UserScript==
@@ -1255,11 +1256,32 @@ function generateChapterList(target, attrURL) {
  */
 let sites = {
 	/**
+	 * MangaFox (Alt Domain)
+	 * @type {SiteObject}
+	 */
+	'mangafox.la' : extendSite({
+		preInit : function(callback) {
+			let ms = sites['mangafox.me'];
+			this.setObjVars      = ms.setObjVars;
+			this.stylize         = ms.stylize;
+			this.preSetupTopBar  = ms.preSetupTopBar;
+			this.postSetupTopBar = ms.postSetupTopBar;
+			this.preSetupViewer  = ms.preSetupViewer;
+			this.setupViewerContainer = ms.setupViewerContainer;
+
+			this.site = 'mangafox.me';
+
+			callback();
+		}
+	}),
+	/**
 	 * MangaFox
 	 * @type {SiteObject}
 	 */
 	'mangafox.me' : extendSite({
 		setObjVars : function () {
+			this.tld = location.hostname.split('.').pop();
+
 			this.segments    = window.location.pathname.split( '/' );
 
 			this.title       = this.segments[2];
@@ -1267,8 +1289,8 @@ let sites = {
 
 			this.page_count  = $('#top_bar').find('.prev_page + div').text().trim().replace(/^[\s\S]*of ([0-9]+)$/, '$1');
 
-			this.title_url   = 'https://mangafox.me/manga/'+this.title+'/';
-			this.chapter_url = '//mangafox.me/manga/'+this.title+'/'+this.chapter+'/';
+			this.title_url   = 'http://mangafox.'+this.tld+'/manga/'+this.title+'/';
+			this.chapter_url = '//mangafox.'+this.tld+'/manga/'+this.title+'/'+this.chapter+'/';
 
 			this.chapterListCurrent = this.chapter_url+'1.html';
 			this.chapterList        = {}; //This is set via preSetupTopbar
@@ -1278,7 +1300,7 @@ let sites = {
 			this.viewerRegex            = /^[\s\S]*(<div class="read_img">[\s\S]*<\/div>)[\s\S]*<\/div>[\s\S]*<div id="shares"[\s\S]*$/;
 			// this.viewerCustomImageList  = []; //This is (possibly) set below.
 
-			this.searchURLFormat = 'https://mangafox.me/search.php?advopts=1&name={%SEARCH%}';
+			this.searchURLFormat = 'http://mangafox.'+this.tld+'/search.php?advopts=1&name={%SEARCH%}';
 
 			this.currentPage = parseInt(this.segments.slice(-1)[0].replace(/^([0-9]+).*/, '$1'));
 
@@ -1297,6 +1319,8 @@ let sites = {
 
 			//Float title in the header to the right. This just looks nicer and is a bit easier to read.
 			tool.find('> #series > strong:last').css('float', 'right');
+
+			$('#left-skyscraper, #right-skyscraper').remove();
 		},
 		preSetupTopBar : function(callback) {
 			let _this = this;
@@ -1305,7 +1329,7 @@ let sites = {
 			//Because of this, we can't use the inline chapter list as a source, and instead we need to check the manga page.
 			//We can't CSRF to the subdomain for some reason, so we need to use a GM function here...
 			GM_xmlhttpRequest({
-				url     : _this.title_url.replace('mangafox.me', 'm.mangafox.me'),
+				url     : _this.title_url.replace('mangafox.'+_this.tld, 'm.mangafox.'+_this.tld),
 				method  : 'GET',
 				onload  : function(response) {
 					let data = response.responseText;
@@ -1317,7 +1341,7 @@ let sites = {
 						let chapterTitle     = $('+ span.title', this).text().trim(),
 						    url              = $(this).attr('href').replace(/^(.*\/)(?:[0-9]+\.html)?$/, '$1'); //Remove trailing page number
 
-						url = url.replace('m.mangafox.me/manga/', 'mangafox.me/manga/');
+						url = url.replace('m.mangafox.'+_this.tld+'/manga/', 'mangafox.'+_this.tld+'/manga/');
 						_this.chapterList[url+'1.html'] = url.replace(/^.*\/manga\/[^/]+\/(?:v(.*?)\/)?c(.*?)\/$/, 'Vol.$1 Ch.$2')
 							.replace(/^Vol\. /, '') + (chapterTitle !== '' ? ': ' + chapterTitle : '');
 					});
@@ -1369,7 +1393,7 @@ let sites = {
 
 			//We can't CSRF to the subdomain for some reason, so we need to use a GM function here...
 			GM_xmlhttpRequest({
-				url     : 'https:'+_this.chapter_url.replace('mangafox.me/manga', 'm.mangafox.me/roll_manga'),
+				url     : 'http:'+_this.chapter_url.replace('mangafox.'+_this.tld+'/manga', 'm.mangafox.'+_this.tld+'/roll_manga'),
 				method  : 'GET',
 				onload  : function(response) {
 					let data      = response.responseText,
@@ -2314,6 +2338,9 @@ let sites = {
 			if(this.segments[3]) {
 				this.currentPage = parseInt(this.segments[3]);
 			}
+		},
+		stylize : function() {
+
 		},
 		preSetupViewer : function(callback) {
 			$('.content').replaceWith($('<div/>', {id: 'viewer'})); //Set base viewer div